requirements.txt
django-statsd-mozilla==0.3.16
suds-jurko==0.6
requests==2.9.1
